parametarized pipeline

+++ Jenkinsfile
@@ -1,5 +1,10 @@
 pipeline {
     agent any
+    parameters {
+        string(name: 'Git Branch',defaultValue:'main')
+        choice(name: 'Terraform Actions', choices: ['Plan', 'Apply','Destroy'], description: 'Terraform Actions')
+        
+    }
 
     environment {
         AWS_ACCESS_KEY_ID     = credentials('AccessKey')   // Jenkins credential ID
@@ -12,7 +17,18 @@ pipeline {
         stage('Checkout Code') {
             steps {
                 script {
-                    checkout scm
+                    try {
+                // Attempt to checkout the specified branch
+                git branch: "${params.BRANCH_NAME}",
+                    url: 'https://github.com/sanashaikh123/terraform-repo-2nd.git',
+                    
+            } catch (Exception e) {
+                // Handle the error, e.g., notify the user or fallback to a default branch
+                echo "Branch '${params.BRANCH_NAME}' not found. Please verify the branch name."
+                currentBuild.result = 'FAILURE'
+                error("Stopping pipeline due to invalid branch.")
+            }
+        }
                 }
             }
         }
@@ -27,36 +43,29 @@ pipeline {
             }
         }
 
-        stage('Terraform Plan') {
+        stage('Terraform Actions') {
             steps {
                 script {
-                    sh 'terraform plan -out=tfplan'
+                    if (params.Actions == "Plan"){
+                        sh 'terraform plan -out=tfplan'
+                    }else if (params.Actions == 'Apply'){
+                        sh 'terraform apply -auto-approve tfplan'
+                    }else if (params.Actions == 'Apply'){
+                        sh 'terraform destroy -auto-approve'
+                    }
+
+                    
                 }
             }
         }
 
-        stage('Terraform Apply') {
-            steps {
-                //input message: "Do you want to apply the Terraform changes?", ok: "Apply"
-                script {
-                    sh 'terraform apply --auto-approve tfplan'
-                }
-            }
-        }
-        stage('Terraform Destroy') {
-            steps {
-                //input message: "Do you want to apply the Terraform changes?", ok: "Apply"
-                Destroy(){
-                
-                    sh 'terraform destroy -auto-approve'
-                }
-            }
-        }
+        
     }
 
     post {
         always {
             echo "Pipeline execution completed"
+            archiveArtifacts artifacts: 'tfplan'
         }
         success {
             echo "Infrastructure provisioned successfully!"
